# Optimized std.datetime.date.Date.fromSimpleString