@JackStouffer 2018-01-25
Forgot that `to!int`/`to!uint` is slightly faster