@wilzbach 2018-01-25
Once https://github.com/dlang/phobos/pull/6070 is merged, CircleCi can be restarted here.